FindExecutable returns a value greater then 32 on success

+++ Code/Commando/WebBrowser.cpp
@@ -1080,12 +1080,12 @@ bool WebBrowser::LaunchExternal(const char* url)
 	// Find the executable that can launch this file
 	char exeName[MAX_PATH];
 	HINSTANCE hInst = FindExecutableA(filename, NULL, exeName);
-	WWASSERT(((int)hInst > 32) && "Unable to find executable that will display HTML files.");
+	WWASSERT(((uintptr_t)hInst > 32) && "Unable to find executable that will display HTML files.");
 
 	// Delete temporary file
 	DeleteFileA(filename);
 
-	if ((int)hInst <= 32)
+	if ((uintptr_t)hInst <= 32)
 		{
 		return false;
 		}
